move comm drivers to u8g_glue.c
